split hdfs
