+++ python/paddle/fluid/tests/unittests/test_auto_checkpoint1.py
@@ -57,7 +57,7 @@ class AutoCheckpointTest1(AutoCheckPointACLBase):
         os.environ.update(proc_env)
 
     def test_corner_epoch_no(self):
-        self._test_core_epoch_no(0)
+        self._test_corner_epoch_no(0)
 
 
 if __name__ == '__main__':
